pipelight - fix button styles
