+++ pipelight/src/components/Styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create(
+  {
+    buttonStyle: {
+      backgroundColor: '#a8b3bc',
+      marginLeft: 15,
+    },
+    tabButtonStyle: {
+      borderRadius: 10,
+      borderWidth: 1,
+      textTransform: 'lowercase',
+    }
+  }
+)
+
+export default styles;
